refactor: test 로직 custom hooks로 변경
