--- src/lib/helpers/context.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Context = React.createContext({});
